job status functionality
added a database column called Status, containing the status of the job application. The user inputs this when submitting the url, and can edit it from the edit_job page. The view_jobs page now also colors each job posting based on the status, to help the user more easily find keep track of the status of each job application (with a legend to show what each color represents). More work needed to improve the visual appeal of app.